mxnet/requirements.txt
uwsgi
falcon==2.0.0
gunicorn==20.0.4
jsonschema==3.2.0
mxnet-mkl==1.5.1.post0
keras-mxnet==2.2.4.2
## The following requirements were added by pip freeze:
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
graphviz==0.8.4
h5py==2.10.0
idna==2.8
importlib-metadata==1.4.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
more-itertools==8.1.0
numpy==1.18.1
pyrsistent==0.15.7
PyYAML==5.3
requests==2.22.0
scipy==1.4.1
six==1.14.0
urllib3==1.25.7
zipp==1.0.0
